requirements.txt
requests
openai
python-dotenv
pyTelegramBotAPI
Pillow
